+++ CHANGELOG.md
@@ -10,11 +10,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Copyright and contributing documentation (#207)
 - Errors when running code now include the name of the file in which the error occurred (#239)
 - Sentry integration (#252)
+- Ability to download a project (#255)
+- FileMenu dropdown(#139)
 
 ### Changed
 
 - Switch wrapping in the editor to horizontal scrolling for long lines (#242)
 - Replaced hard-coded text with translatable strings (#253)
+- Launch rename modal via redux (#139)
 
 ### Fixed
 - Allow users to save their code as anew project from the root page when logged in (#259)